refactor(189): move BlockAccumulator to core crate
Move BlockAccumulator, OpenBlock, StreamFinalize, and finalize_blocks
from the adapters crate into swink-agent core so both adapters and
local-llm can share the same streaming event-assembly state machine.

Adapters' block_accumulator.rs and finalize.rs become thin re-export
modules — no public API change for existing adapter consumers.

Part 1 of the #189 local-llm migration split (PR #258 exceeded the
1500-line diff limit). Part 2 will migrate local-llm's StreamState
to use the core BlockAccumulator.
